extracted validateBulkDeleteSubjectSize function
